Remove unused finetune and freeze modules
